Fix pdf download for records with single-page pdfs
- legacy pdf download uses same creation code as creation in message
queue

+++ goobi-viewer-core/src/main/java/io/goobi/viewer/model/job/download/PdfDownloadJob.java
@@ -125,7 +125,10 @@ public class PdfDownloadJob extends DownloadJob {
 
     public void create(Dataset work)
             throws IOException, ContentLibException, PresentationException {
-        createLock();
+        if (!createLock()) {
+            // Another thread/process is already generating this file; do nothing and let the caller wait/retry.
+            return;
+        }
         Files.deleteIfExists(getTempPath());
         try (FileOutputStream fos = new FileOutputStream(getTempPath().toFile())) {
             MetsPdfRequest request = createPdfRequest(work,
@@ -194,7 +197,7 @@ public class PdfDownloadJob extends DownloadJob {
         return sb.toString();
     }
 
-    private static MetsPdfRequest createPdfRequest(Dataset work, Optional<String> divId, boolean usePdfSource, String configVariant)
+    public static MetsPdfRequest createPdfRequest(Dataset work, Optional<String> divId, boolean usePdfSource, String configVariant)
             throws URISyntaxException {
 
         boolean usePdfFiles = usePdfSource;
